Move authorization resource to `:shared:security` module.
