+++ controller/src/test/resources/META-INF/services/io.pravega.auth.AuthHandler
@@ -7,4 +7,4 @@
 #
 #     http://www.apache.org/licenses/LICENSE-2.0
 #
-io.pravega.controller.server.security.auth.handler.TestAuthHandler
+io.pravega.controller.auth.TestAuthHandler
