+++ shared/security/src/main/java/io/pravega/shared/security/auth/AuthorizationResourceImpl.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.controller.server.security.auth;
+package io.pravega.shared.security.auth;
 
 import io.pravega.common.Exceptions;
 import io.pravega.shared.NameUtils;
